chatbot/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Django==5.1.3
fastapi==0.115.0
frozenlist==1.4.1
gigachain==0.2.16
gigachain-cli==0.0.30.post1
gigachain-community==0.2.16.post2
gigachain-core==0.2.38.post2
gigachain-text-splitters==0.2.4
gigachat==0.1.35
gigaserve==0.2.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langfuse==2.50.2
langsmith==0.1.125
libcst==1.4.0
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyproject-toml==0.0.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.2
sse-starlette==1.8.2
starlette==0.38.5
tenacity==8.5.0
toml==0.10.2
tomlkit==0.12.5
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.11.1
